# specs/SPEC/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InfinityMix API",
    "version": "1.0.0",
    "description": "API for InfinityMix, an AI-powered mashup generation platform."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "paths": {
...
    "/mashups/generate": {
      "post": {
        "tags": [
          "mashups"
        ],
        "summary": "Trigger AI mashup generation",
        "operationId": "generateMashup",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MashupGenerateRequest"
              },
              "example": {
                "input_file_ids": [
                  "f1e2d3c4-b5a6-7890-1234-567890abcdef",
                  "a9b8c7d6-e5f4-3210-9876-543210fedcba"
                ],
                "duration_preset": "2_minutes"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Mashup generation initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MashupResponse"
                },
                "example": {
                  "id": "m1a2s3h4-u5p6-7890-1234-567890abcdef",
                  "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                  "duration_seconds": 120,
                  "status": "queued",
                  "output_path": null,
                  "output_format": "mp3",
                  "generation_time_ms": null,
                  "created_at": "2025-11-28T12:05:00Z",
                  "updated_at": "2025-11-28T12:05:00Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/mashups/{mashupId}": {
      "get": {
        "tags": [
          "mashups"
        ],
        "summary": "Get details of a specific mashup",
        "operationId": "getMashupDetails",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mashupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the mashup to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Mashup details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MashupResponse"
                },
                "example": {
                  "id": "m1a2s3h4-u5p6-7890-1234-567890abcdef",
                  "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                  "duration_seconds": 120,
                  "status": "completed",
                  "output_path": "https://storage.example.com/mashups/m1a2s3h4-u5p6.mp3",
                  "output_format": "mp3",
                  "generation_time_ms": 75000,
                  "created_at": "2025-11-28T12:05:00Z",
                  "updated_at": "2025-11-28T12:06:15Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "mashups"
        ],
        "summary": "Delete a specific mashup",
        "operationId": "deleteMashup",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mashupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the mashup to delete."
          }
        ],
        "responses": {
          "204": {
            "description": "Mashup deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/mashups/{mashupId}/download": {
      "get": {
        "tags": [
          "mashups"
        ],
        "summary": "Download a generated mashup file",
        "operationId": "downloadMashup",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mashupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the mashup to download."
          }
        ],
        "responses": {
          "200": {
            "description": "Mashup file downloaded successfully",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Mashup not yet ready for download (still processing or failed)"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/mashups/{mashupId}/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit feedback for a generated mashup",
        "operationId": "submitMashupFeedback",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mashupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the mashup to provide feedback for."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              },
              "example": {
                "rating": 5,
                "comments": "Absolutely loved this mashup! The transitions were seamless."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                },
                "example": {
                  "id": "f3e2d1c0-b9a8-7654-3210-fedcba987654",
                  "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                  "mashup_id": "m1a2s3h4-u5p6-7890-1234-567890abcdef",
                  "rating": 5,
                  "comments": "Absolutely loved this mashup! The transitions were seamless.",
                  "created_at": "2025-11-28T12:10:00Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Feedback already submitted for this mashup by this user."
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MashupGenerateRequest": {
        "type": "object",
        "required": [
          "input_file_ids",
          "duration_preset"
        ],
        "properties": {
          "input_file_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 2,
            "maxItems": 5,
            "description": "List of IDs of uploaded audio files to use for the mashup. (MVP: 2-5 files)"
          },
          "duration_preset": {
            "type": "string",
            "enum": [
              "1_minute",
              "2_minutes",
              "3_minutes"
            ],
            "description": "Predefined duration for the generated mashup. (MVP: fixed presets)"
          }
        }
      },
      "MashupResponse": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "duration_seconds",
          "status",
          "output_format",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the mashup."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who generated the mashup."
          },
          "duration_seconds": {
            "type": "integer",
            "description": "Target duration of the mashup in seconds."
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed"
            ],
            "description": "Current status of the mashup generation."
          },
          "output_path": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL to the generated mashup file (available if status is 'completed')."
          },
          "output_format": {
            "type": "string",
            "description": "Format of the output file (e.g., 'mp3')."
          },
          "generation_time_ms": {
            "type": "integer",
            "nullable": true,
            "description": "Time taken for generation in milliseconds (available if status is 'completed')."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the generation request was made."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the mashup status was last updated."
          }
        }
      },
...
      "FeedbackRequest": {
        "type": "object",
        "required": [
          "rating"
        ],
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "User's satisfaction rating (1-5 stars)."
          },
          "comments": {
            "type": "string",
            "nullable": true,
            "maxLength": 500,
            "description": "Optional text comments from the user."
          }
        }
      },
      "FeedbackResponse": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "mashup_id",
          "rating",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the feedback entry."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who provided the feedback."
          },
          "mashup_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the mashup being rated."
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "User's satisfaction rating (1-5 stars)."
          },
          "comments": {
            "type": "string",
            "nullable": true,
            "description": "Optional text comments from the user."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the feedback was submitted."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "A human-readable error message."
          },
          "code": {
            "type": "string",
            "description": "An optional, application-specific error code."
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input provided (e.g., missing fields, invalid format).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Invalid email format."
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Invalid email or password."
            }
          }
        }
      },
      "Forbidden": {
        "description": "User does not have permission to access this resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "You do not have permission to delete this mashup."
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Mashup with ID 'xyz' not found."
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "An internal server error occurred. Please try again later."
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}